perbaiki delete gambar

+++ App/CRUD/Question/delete.php
@@ -25,7 +25,9 @@ if(isset($_POST['delete'])){
     // hapus balasan dan gambar jawaban
     foreach($answers as $answer){
         delete_byId("replies", "answer_id", $answer['id']);
-        delete_file("img/", $answer['image']);
+        if( $answer['image'] != "no image" ){
+            delete_file("img/", $answer['image']);
+        }
     }
 
     // hapus jawaban
@@ -37,7 +39,9 @@ if(isset($_POST['delete'])){
     // hapus
     if(delete_byId("questions", "id", $id) > 0){
 
-        delete_file("img/", $question['image']);
+        if( $question['image'] != "no image" ){
+            delete_file("img/", $question['image']);
+        }
         echo " 
                 <script> 
                     alert('data berhasil dihapus');
